third_party/blink/web_tests/external/wpt/html/canvas/offscreen/layers/2d.layer.malformed-operations.transferToImageBitmap.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.layer.malformed-operations.transferToImageBitmap</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.layer.malformed-operations.transferToImageBitmap</h1>
<p class="desc">Check that exceptions are thrown for operations that are malformed while layers are open.</p>


<script>
test(t => {
  var canvas = new OffscreenCanvas(200, 200);
  var ctx = canvas.getContext('2d');

  // Shouldn't throw on its own.
  canvas.transferToImageBitmap();
  // Make sure the exception isn't caused by calling the function twice.
  canvas.transferToImageBitmap();
  // Calling again inside a layer should throw.
  ctx.beginLayer();
  assert_throws_dom("InvalidStateError",
                    () => canvas.transferToImageBitmap());

}, "Check that exceptions are thrown for operations that are malformed while layers are open.");
</script>
